fix chapter numbering
Section numbers in output were off by one because the generated
document starte off with the preface as section number 1.

By enabling pandoc's attributes extension, we can make the Preface
section unnumbered.  Thus the introduction becomes the first chapter,
and the remaining chapters follow the numbering of the input files.